+++ hexa/pipelines/tests/test_models.py
@@ -1,6 +1,7 @@
 from django.core import mail
 
 from hexa.core.test import TestCase
+from hexa.pipeline_templates.models import PipelineTemplateVersion
 from hexa.pipelines.models import (
     Pipeline,
     PipelineNotificationLevel,
@@ -85,6 +86,13 @@ class PipelineTest(TestCase):
             countries=[{"code": "AL"}],
         )
 
+        cls.WORKSPACE2 = Workspace.objects.create_if_has_perm(
+            cls.USER_ADMIN,
+            name="Sandbox2",
+            description="This is a sandbox workspace ",
+            countries=[{"code": "AL"}],
+        )
+
         cls.PIPELINE = Pipeline.objects.create(
             workspace=cls.WORKSPACE,
             name="Test pipeline",
@@ -92,7 +100,14 @@ class PipelineTest(TestCase):
             description="This is a test pipeline",
         )
         cls.PIPELINE.upload_new_version(
-            cls.USER_ADMIN, zipfile=b"", parameters=[], name="Version"
+            cls.USER_ADMIN,
+            zipfile=b"",
+            name="Version",
+            parameters=[
+                {"code": "param_1", "name": "Param 1", "default": 123, "type": "int"},
+                {"code": "param_2", "name": "Param 2", "default": 456, "type": "int"},
+            ],
+            config={"param_1": 1234, "param_2": 4567},
         )
 
     def test_mail_run_recipients_mail_not_sent(self):
@@ -291,7 +306,7 @@ class PipelineTest(TestCase):
             config=None,
         )
         self.assertEqual(
-            {"param_1": 45, "param_2": 46},
+            {"param_1": 43, "param_2": 46},
             pipeline.last_version.config,
             "Config from previous version with a change of default values",
         )
@@ -302,7 +317,6 @@ class PipelineTest(TestCase):
             name=template_name,
             code="test_code",
             description="Some description",
-            config={"key": "value"},
         )
         self.assertIsNotNone(template)
         self.assertEqual(self.PIPELINE.template.name, template_name)
@@ -310,7 +324,6 @@ class PipelineTest(TestCase):
             name="SOME RANDOM NAME",
             code="test_code",
             description="Some description",
-            config={"key": "value"},
         )
         self.assertEqual(
             self.PIPELINE.template.name, template_name
@@ -321,7 +334,6 @@ class PipelineTest(TestCase):
             name="Test Template",
             code="test_code",
             description="Some description",
-            config={"key": "value"},
         )
         template_version = template.create_version(
             self.PIPELINE.last_version, changelog="First version"
@@ -331,22 +343,56 @@ class PipelineTest(TestCase):
             self.USER_ADMIN, self.WORKSPACE2
         )
         created_pipeline = created_pipeline_version.pipeline
-        self.assertEqual(len(created_pipeline.new_template_versions), 0)
+        self.assertFalse(created_pipeline.has_new_template_versions)
 
         for i in range(2, 5):
             self.PIPELINE.upload_new_version(
                 user=self.USER_ADMIN,
                 zipfile=b"",
-                parameters=[],
+                parameters=[
+                    {
+                        "code": "param_1",
+                        "name": "Param 1",
+                        "default": 444,
+                        "type": "int",
+                    },
+                    {
+                        "code": "param_3",
+                        "name": "Param 3",
+                        "default": 666,
+                        "type": "int",
+                    },
+                    {"code": "param_4", "name": "Param 4", "type": "dhis2"},
+                ],
                 name=f"Version {i}",
-                config={},
+                config={"param_1": 888, "param_3": 999, "param_4": "dhis2"},
             )
             template.create_version(
                 self.PIPELINE.last_version, changelog=f"Changelog {i}"
             )
 
-        self.assertEqual(len(created_pipeline.new_template_versions), 3)
+        self.assertEqual(
+            PipelineTemplateVersion.objects.get_updates_for(created_pipeline).count(), 3
+        )
+
+        created_pipeline.upload_new_version(
+            self.USER_ADMIN,
+            zipfile=b"",
+            name="Version",
+            parameters=[
+                {"code": "param_1", "name": "Param 1", "default": 123, "type": "int"},
+                {"code": "param_5", "name": "Param 5", "default": 456, "type": "int"},
+            ],
+            config={"param_1": 1234, "param_5": 4567},
+        )
+
+        self.assertEqual(
+            PipelineTemplateVersion.objects.get_updates_for(created_pipeline).count(), 3
+        )
 
         template.upgrade_pipeline(self.USER_ADMIN, created_pipeline)
 
-        self.assertEqual(len(created_pipeline.new_template_versions), 0)
+        self.assertFalse(created_pipeline.has_new_template_versions)
+        self.assertEqual(
+            created_pipeline.last_version.config, {"param_1": 1234, "param_3": 666}
+        )